favourites-migration/requirements.txt
google-cloud-datastore
psycopg2